+++ src/renderer/component/fileCard/view.jsx
@@ -91,18 +91,16 @@ class FileCard extends React.PureComponent<Props> {
         onContextMenu={handleContextMenu}
       >
         <CardMedia thumbnail={thumbnail} />
-        <div className="card__title-identity">
-          <div className="card__title--small card__title--file-card">
-            <TruncatedText text={title} lines={2} />
-          </div>
-          <div className="card__subtitle">
-            {pending ? <div>Pending...</div> : <UriIndicator uri={uri} link />}
-          </div>
-          <div className="card__file-properties">
-            {showPrice && <FilePrice hideFree uri={uri} />}
-            {isRewardContent && <Icon iconColor="red" icon={icons.FEATURED} />}
-            {fileInfo && <Icon icon={icons.LOCAL} />}
-          </div>
+        <div className="card__title card__title--file-card">
+          <TruncatedText text={title} lines={2} />
+        </div>
+        <div className="card__subtitle">
+          {pending ? <div>Pending...</div> : <UriIndicator uri={uri} link />}
+        </div>
+        <div className="card__file-properties">
+          {showPrice && <FilePrice hideFree uri={uri} />}
+          {isRewardContent && <Icon iconColor="red" icon={icons.FEATURED} />}
+          {fileInfo && <Icon icon={icons.LOCAL} />}
         </div>
       </section>
     );